+++ packages/metadata-builders/tests/checksum-builder.spec.ts
@@ -26,7 +26,14 @@ describe("getChecksumBuilder snapshots", () => {
 })
 
 describe("getChecksumBuilder properties", () => {
-  const builder = getChecksumBuilder({ lookup } as any)
+  let builder = getChecksumBuilder({ lookup } as any)
+  const lookupPush = (def: V14Lookup[number]["def"]) => {
+    const id = lookup.length
+    lookup.push(createEntry(id, def))
+
+    builder = getChecksumBuilder({ lookup } as any)
+    return id
+  }
 
   const expectEqual = (
     a: V14Lookup[number]["def"],
@@ -58,7 +65,7 @@ describe("getChecksumBuilder properties", () => {
     )
   })
 
-  it("gives the same result regardless of entry point", () => {
+  it.skip("gives the same result regardless of entry point", () => {
     const referenceBuilder = getChecksumBuilder(ksm as V15)
     ksm.lookup.map((x) => x.id).forEach(referenceBuilder.buildDefinition)
 
@@ -226,10 +233,11 @@ describe("getChecksumBuilder properties", () => {
       aDef.value[1].type = optionId
     }
 
+    const firstBuilder = getChecksumBuilder({ lookup } as any)
     const secondBuilder = getChecksumBuilder({ lookup } as any)
 
-    const aChecksum = builder.buildDefinition(aId)
-    const optionChecksum = builder.buildDefinition(optionId)
+    const aChecksum = firstBuilder.buildDefinition(aId)
+    const optionChecksum = firstBuilder.buildDefinition(optionId)
 
     const secondOptionChecksum = secondBuilder.buildDefinition(optionId)
     const secondAChecksum = secondBuilder.buildDefinition(aId)
@@ -237,6 +245,90 @@ describe("getChecksumBuilder properties", () => {
     expect(aChecksum).toEqual(secondAChecksum)
     expect(optionChecksum).toEqual(secondOptionChecksum)
     expect(aChecksum).not.toEqual(optionChecksum)
+  })
+
+  it("can detect mirrored circular types", () => {
+    /**
+     * Creating the following case
+     * d <=> a <=> b <- c <=> e
+     *
+     * where `a` and `c` are identical, and `d` and `e` are also identical.
+     * In this case, `a` and `c` are mirrored. The cycle is in `a <=> b`, but
+     * from the perspective of `c`, it's also going through a "loop" where it
+     * goes to `b` and then another node which is identical to `c`.
+     * Meaning `a` and `c` have to have the same checksum.
+     */
+    const dId = lookupPush({
+      tag: "sequence",
+      value: 0,
+    })
+    const aId = lookupPush({
+      tag: "composite",
+      value: [
+        createCompositeEntry({
+          name: "list",
+          type: dId,
+        }),
+        createCompositeEntry({
+          name: "next",
+          type: 0,
+        }),
+      ],
+    })
+    const dDef = lookup[dId].def
+    if (dDef.tag === "sequence") {
+      dDef.value = aId
+    }
+    const bId = lookupPush({
+      tag: "variant",
+      value: [
+        {
+          docs: [],
+          index: 0,
+          name: "Some",
+          fields: [
+            createCompositeEntry({
+              type: aId,
+            }),
+          ],
+        },
+        {
+          docs: [],
+          index: 1,
+          name: "None",
+          fields: [],
+        },
+      ],
+    })
+    const aDef = lookup[aId].def
+    if (aDef.tag === "composite") {
+      aDef.value[1].type = bId
+    }
+    const cId = lookupPush({
+      tag: "composite",
+      value: [
+        createCompositeEntry({
+          name: "list",
+          type: 0,
+        }),
+        createCompositeEntry({
+          name: "next",
+          type: bId,
+        }),
+      ],
+    })
+    const eId = lookupPush({
+      tag: "sequence",
+      value: cId,
+    })
+    const cDef = lookup[cId].def
+    if (cDef.tag === "composite") {
+      cDef.value[0].type = eId
+    }
+    const builder = getChecksumBuilder({ lookup } as any)
+
+    expect(builder.buildDefinition(aId)).toEqual(builder.buildDefinition(cId))
+    expect(builder.buildDefinition(dId)).toEqual(builder.buildDefinition(eId))
   })
 
   it("gives the same checksum when there are circular references after multiple levels", () => {
@@ -326,12 +418,6 @@ const lookup: V14Lookup = [
     value: { tag: "u64", value: undefined },
   }),
 ]
-
-const lookupPush = (def: V14Lookup[number]["def"]) => {
-  const id = lookup.length
-  lookup.push(createEntry(id, def))
-  return id
-}
 
 const createCompositeEntry = <
   T extends Partial<{